Move errors definitions in a separate file
