+++ figures-counter/includes/figures-counter.h
@@ -1,8 +1,9 @@
 ﻿#pragma once
 
+#include "error.h"
+
 #include <algorithm>
 #include <cassert>
-#include <format>
 #include <istream>
 #include <limits.h>
 #include <memory>
@@ -11,95 +12,6 @@
 #include <vector>
 
 namespace figures_counter {
-
-class figures_counter_exception : public std::invalid_argument {
-public:
-	figures_counter_exception(const std::string& what) :
-		invalid_argument(std::format("figures counter error: {}", what)) {
-	}
-};
-
-class bad_input : public figures_counter_exception {
-public:
-	bad_input(const std::string& what) :
-		figures_counter_exception(std::format("bad input: {}", what)) {
-	}
-};
-
-class row_length_error : public bad_input {
-public:
-	row_length_error(unsigned row, unsigned length, unsigned expected_length) :
-		bad_input(std::format("row {} with length {}, expected length {}", row, length, expected_length)),
-		_row(row),
-		_length(length),
-		_expected_length(expected_length) {
-	}
-
-	unsigned row() const noexcept {
-		return _row;
-	}
-
-	unsigned length() const noexcept {
-		return _length;
-	}
-
-	unsigned expected_length() const noexcept {
-		return _expected_length;
-	}
-
-private:
-	unsigned _row;
-	unsigned _length;
-	unsigned _expected_length;
-};
-
-class row_count_error : public bad_input {
-public:
-	row_count_error(unsigned count, unsigned expected_count) :
-		bad_input(std::format("row count is {}, expected count {}", count, expected_count)),
-		_count(count),
-		_expected_count(expected_count) {
-	}
-
-	unsigned count() const noexcept {
-		return _count;
-	}
-
-	unsigned expected_count() const noexcept {
-		return _expected_count;
-	}
-
-private:
-	unsigned _count;
-	unsigned _expected_count;
-};
-
-class unrecognized_symbol : public bad_input {
-public:
-	unrecognized_symbol(unsigned row, unsigned col, char symbol) :
-		bad_input(std::format("received unrecognized char \"{}\" at [row = {}, col = {}]", symbol, row, col)),
-		_row(row),
-		_col(col),
-		_symbol(symbol) {
-	}
-
-	unsigned row() const noexcept {
-		return _row;
-	}
-
-	unsigned col() const noexcept {
-		return _col;
-	}
-
-	char symbol() const noexcept {
-		return _symbol;
-	}
-
-private:
-	unsigned _row;
-	unsigned _col;
-	char _symbol;
-};
 
 /**
  * @brief The accepted input values for a cell of the input table
@@ -137,13 +49,13 @@ constexpr unsigned EMPTY_FIGURE_ID = UINT_MAX;
 
 int count_figures(int rows, int cols, std::istream& data) {
 	if (rows <= 0) {
-		throw bad_input("rows must be positive");
+		throw error::bad_input("rows must be positive");
 	}
 	if (cols <= 0) {
-		throw bad_input("cols must be positive");
+		throw error::bad_input("cols must be positive");
 	}
 	if (!data) {
-		throw bad_input("data stream can not be read");
+		throw error::bad_input("data stream can not be read");
 	}
 
 	std::string buffer;
@@ -167,10 +79,10 @@ int count_figures(int rows, int cols, std::istream& data) {
 	for (size_t row = 0; row < rows; ++row) {
 		load_next_row();
 		if (!data) {
-			throw row_count_error(row, rows);
+			throw error::row_count_error(row, rows);
 		}
 		if (buffer.size() != cols) {
-			throw row_length_error(row, buffer.size() - 1, cols);
+			throw error::row_length_error(row, buffer.size() - 1, cols);
 		}
 
 		// Cache the left cell in to skip check whether the current cell is 0
@@ -199,7 +111,7 @@ int count_figures(int rows, int cols, std::istream& data) {
 					current_row.push_back(left_cell_figure = sorted.first);
 				}
 			} else [[unlikely]] {
-				throw unrecognized_symbol(row, col, cell);
+				throw error::unrecognized_symbol(row, col, cell);
 			}
 		}
 	}
